feat: basic clean of code
