+++ package.json
@@ -18,7 +18,6 @@
     "@types/react-router": "^4.0.22",
     "@types/react-router-dom": "^4.2.4",
     "@types/react-router-redux": "^5.0.12",
-    "autoprefixer": "^8.1.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.4",
     "babel-plugin-import": "^1.6.7",
@@ -26,7 +25,6 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "css-loader": "^0.28.11",
-    "express": "^4.16.3",
     "file-loader": "^1.1.11",
     "fs-extra": "^5.0.0",
     "html-webpack-plugin": "^3.1.0",
@@ -35,9 +33,6 @@
     "less-loader": "^4.1.0",
     "mini-css-extract-plugin": "^0.4.0",
     "optimize-css-assets-webpack-plugin": "^4.0.0",
-    "parcel-bundler": "^1.6.2",
-    "postcss": "^6.0.19",
-    "postcss-modules": "^1.1.0",
     "prettier": "^1.11.1",
     "react-dev-utils": "^5.0.0",
     "style-loader": "^0.20.3",
